Updated udp_client.py and events
the udp_client.py file is now independent of the program and can listen for messages from the traffic generator. Moreover the traffic generator is up and running giving messages to the client file. No points or anything is has changed with the database just getting the events to display correctly. The udp_server.py and pythonTrafficGenerator.py cannot run at the same time as they will try to bind to the same server address and one will fail.